feat: tested and created methods to add response(s) to bucket(s)

+++ src/main/java/eventprocessing/models/InitialResponseList.java
@@ -1,6 +1,5 @@
 package eventprocessing.models;
 
-import java.math.BigInteger;
 import java.util.Comparator;
 import java.util.Optional;
 
@@ -10,7 +9,7 @@ public class InitialResponseList extends ResponseList {
         super();
     }
 
-    public BigInteger getEarliestTimestamp() {
+    public long getEarliestTimestamp() {
         Optional<Response> earliestResponse = getResponses().stream()
                 .min(Comparator.comparing(Response::getTimestamp));
