+++ src/main/java/eventprocessing/models/Message.java
@@ -1,12 +1,10 @@
 package eventprocessing.models;
 
-import java.math.BigInteger;
-
 public class Message {
     private String locationId;
     private String eventId;
     private double value;
-    private BigInteger timestamp;
+    private long timestamp;
 
     public String getLocationId() {
         return locationId;
@@ -16,7 +14,7 @@ public class Message {
         return value;
     }
 
-    public BigInteger getTimestamp() {
+    public long getTimestamp() {
         return timestamp;
     }
 }
